refactor: use 32-bit floats instead of 64-bit

+++ src/solver.rs
@@ -6,22 +6,22 @@ pub enum SortStrategy {
 }
 
 pub struct Cart {
-    pub items: Vec<f64>,
-    pub total_cost: f64,
+    pub items: Vec<f32>,
+    pub total_cost: f32,
 }
 
 // Return maximum list of items that fit within given budget.
 // Supports multiple methods for sorting the list before greedily pulling items.
-pub fn grab_max_items(wishlist: &[f64], budget: f64, strategy: SortStrategy) -> Cart {
+pub fn grab_max_items(wishlist: &[f32], budget: f32, strategy: SortStrategy) -> Cart {
     let mut sorted_wishlist = wishlist.to_vec();
-    sorted_wishlist.sort_by(f64::total_cmp);
+    sorted_wishlist.sort_by(f32::total_cmp);
 
     if strategy == SortStrategy::Expensive {
         sorted_wishlist.reverse();
     }
 
-    let mut total_cost = 0_f64;
-    let mut cart = Vec::<f64>::new();
+    let mut total_cost = 0_f32;
+    let mut cart = Vec::<f32>::new();
 
     for cost in sorted_wishlist {
         if total_cost + cost <= budget {
